+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- `@reykjavik/webtools/CookieHubConsent`:
+  - feat: Allow passing explicit `undefined` as `accountId` to skip loading
 
 ## 0.2.1